# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema de Gestión Odontológica - API
  version: 0.2.0
  description: API REST del backend. Se autentica con una cookie de sesión
    HttpOnly. Las escrituras requieren el header Origin de la aplicación y el
    token CSRF de la sesión. Los errores siguen el formato ErrorBody.
servers:
  - url: /
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: Cookie de sesión HttpOnly que emite el login o el setup inicial
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Token de la sesión (viene en la respuesta de login y de /auth/me).
        Las escrituras también exigen el header Origin de la app.
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: APPOINTMENT_CONFLICT
            message:
              type: string
              example: El profesional ya tiene un turno en ese horario
            details: {}
          required:
            - code
            - message
      required:
        - error
...
    ClinicalEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entryType:
          type: string
          enum:
            - EVOLUTION
            - CORRECTION
        content:
          type: string
        correctionOfId:
          type:
            - string
            - "null"
          format: uuid
        appointmentId:
          type:
            - string
            - "null"
          format: uuid
        professional:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
          required:
            - id
            - displayName
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - entryType
        - content
        - correctionOfId
        - appointmentId
        - professional
        - createdAt
    ClinicalProfile:
      type: object
      properties:
        current:
          type:
            - object
            - "null"
          properties:
            alerts:
              type:
                - string
                - "null"
            allergies:
              type:
                - string
                - "null"
            medications:
              type:
                - string
                - "null"
            background:
              type:
                - string
                - "null"
            professional:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                displayName:
                  type: string
              required:
                - id
                - displayName
            updatedAt:
              type: string
              format: date-time
          required:
            - alerts
            - allergies
            - medications
            - background
            - professional
            - updatedAt
        versions:
          type: integer
          description: Cantidad de versiones guardadas
      required:
        - current
        - versions
    SaveClinicalProfileInput:
      type: object
      properties:
        alerts:
          type:
            - string
            - "null"
          maxLength: 300
          description: Alertas breves que se muestran en la ficha
        allergies:
          type:
            - string
            - "null"
          maxLength: 1000
        medications:
          type:
            - string
            - "null"
          maxLength: 1000
        background:
          type:
            - string
            - "null"
          maxLength: 4000
      required:
        - alerts
        - allergies
        - medications
        - background
    CreateClinicalEntryInput:
      type: object
      properties:
        content:
          type: string
          minLength: 3
          maxLength: 10000
        appointmentId:
          type:
            - string
            - "null"
          format: uuid
      required:
        - content
    CorrectClinicalEntryInput:
      type: object
      properties:
        content:
          type: string
          minLength: 3
          maxLength: 10000
      required:
        - content
    ClinicalFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        mimeType:
          type: string
          enum:
            - application/pdf
            - image/jpeg
            - image/png
        size:
          type: integer
        category:
          $ref: "#/components/schemas/ClinicalFileCategory"
        description:
          type:
            - string
            - "null"
        clinicalEntryId:
          type:
            - string
            - "null"
          format: uuid
        professional:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
          required:
            - id
            - displayName
        createdAt:
          type: string
          format: date-time
        archived:
          type: boolean
      required:
        - id
        - filename
        - mimeType
        - size
        - category
        - description
        - clinicalEntryId
        - professional
        - createdAt
        - archived
    ClinicalFileCategory:
      type: string
      enum:
        - XRAY
        - STUDY
        - PHOTO
        - CONSENT
        - OTHER
...
paths:
...
  /api/patients/{patientId}/clinical:
    get:
      tags: &a9
        - Historia clínica
      summary: Historia clínica del paciente
      description: Solo los odontólogos acceden. Administración y recepción reciben
        403. Cada apertura de la historia queda registrada en la auditoría.
      security: *a2
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
      responses:
        "200":
          description: Notas, de la más reciente a la más antigua
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClinicalEntry"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags: *a9
      summary: Registra una nota de evolución
      description: "La nota queda a nombre del odontólogo y no se puede editar ni
        borrar: se corrige agregando una corrección."
      security: *a1
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClinicalEntryInput"
      responses:
        "201":
          description: Nota registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalEntry"
        "400":
          description: Los datos enviados no son válidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Paciente archivado, usuario sin perfil profesional o turno que no
            corresponde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/patients/{patientId}/clinical/profile:
    get:
      tags: *a9
      summary: "Perfil clínico vigente: alertas, alergias, medicación y antecedentes"
      description: Solo odontólogos. La consulta queda auditada.
      security: *a2
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
      responses:
        "200":
          description: Perfil clínico
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalProfile"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    put:
      tags: *a9
      summary: Guarda una nueva versión del perfil clínico
      description: Cada cambio crea una versión nueva; las anteriores se conservan.
      security: *a1
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveClinicalProfileInput"
      responses:
        "200":
          description: Perfil clínico actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalProfile"
        "400":
          description: Los datos enviados no son válidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Paciente archivado o usuario sin perfil profesional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/patients/{patientId}/clinical/{entryId}/corrections:
    post:
      tags: *a9
      summary: Agrega una corrección a una nota
      description: La nota original se conserva intacta; la corrección queda vinculada
        a ella.
      security: *a1
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
        - schema:
            type: string
            format: uuid
          required: true
          name: entryId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrectClinicalEntryInput"
      responses:
        "201":
          description: Adenda registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalEntry"
        "400":
          description: Los datos enviados no son válidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Paciente archivado, sin perfil profesional, o la nota ya es una
            corrección
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/patients/{patientId}/files:
    get:
      tags: &a10
        - Archivos clínicos
      summary: Archivos clínicos del paciente
      description: Solo odontólogos. Por defecto no incluye los archivados.
      security: *a2
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
        - schema:
            type: string
            enum:
              - "true"
              - "false"
            default: "false"
          required: false
          name: archived
          in: query
      responses:
        "200":
          description: Archivos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClinicalFile"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags: *a10
      summary: Sube un archivo clínico (PDF, JPG o PNG)
      description: El cuerpo es el contenido binario del archivo. El tipo se verifica
        por su contenido, no por la extensión. El tamaño máximo es configurable
        (20 MB por defecto). El archivo se guarda fuera de cualquier carpeta
        pública.
      security: *a1
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
        - schema:
            type: string
            minLength: 1
            maxLength: 200
          required: true
          name: filename
          in: query
        - schema:
            $ref: "#/components/schemas/ClinicalFileCategory"
          required: true
          name: category
          in: query
        - schema:
            type: string
            maxLength: 300
          required: false
          name: description
          in: query
        - schema:
            type: string
            format: uuid
          required: false
          name: clinicalEntryId
          in: query
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Archivo guardado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalFile"
        "400":
          description: Los datos enviados no son válidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "413":
          description: El archivo supera el tamaño máximo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "415":
          description: Tipo de archivo no permitido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Archivo vacío, paciente archivado, sin perfil profesional o nota
            que no corresponde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/patients/{patientId}/files/{fileId}/content:
    get:
      tags: *a10
      summary: Descarga o visualiza un archivo clínico
      description: Solo a través del backend, con sesión y permiso clínico. Cada
        descarga queda auditada.
      security: *a2
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
        - schema:
            type: string
            format: uuid
          required: true
          name: fileId
          in: path
      responses:
        "200":
          description: Contenido del archivo
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/patients/{patientId}/files/{fileId}/archive:
    post:
      tags: *a10
      summary: Archiva un archivo clínico (no se borra)
      security: *a1
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: patientId
          in: path
        - schema:
            type: string
            format: uuid
          required: true
          name: fileId
          in: path
      responses:
        "200":
          description: Archivo archivado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalFile"
        "401":
          description: Sin sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Sin permiso, o falla la verificación CSRF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: El recurso no existe (o pertenece a otra organización)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...